refactor : dashboard_screen.dart를 상태/순수로직/액션으로 분리
admin/features/dashboard 폴더가 UI, 상태 관리, 순수 로직, 다이얼로그/삭제
액션이 한 파일(751줄)에 섞여 있어 track_bulk_manage 등 다른 feature가
따르는 컨벤션(상태·로직·액션을 별도 파일로 분리)과 어긋났다.

- dashboard_state.dart: 정렬/필터 enum과 Notifier provider
- dashboard_entries.dart: CornerDashboardEntry와 build/filter/sort/format 순수 함수
- dashboard_actions.dart: 코너 추가 다이얼로그, 코너 삭제 액션
- dashboard_screen.dart: 위젯 트리만 남김

부수적으로 frontend/Makefile의 docker-check가 packages/cornermon_api_gen
(자체 pubspec을 가진 별도 패키지)에서 pub get을 하지 않아 컨테이너 안에서
전부 미해결 타입 에러가 나는 기존 버그를 발견해 함께 수정했다.
