+++ frontend/test/admin/features/dashboard/dashboard_screen_test.dart
@@ -1,6 +1,8 @@
 import 'dart:async';
 
+import 'package:cornermon/admin/features/dashboard/dashboard_entries.dart';
 import 'package:cornermon/admin/features/dashboard/dashboard_screen.dart';
+import 'package:cornermon/admin/features/dashboard/dashboard_state.dart';
 import 'package:cornermon/admin/features/track_bulk_manage/track_pin_export_controller.dart';
 import 'package:cornermon/admin/features/track_direct/track_direct_providers.dart';
 import 'package:cornermon/admin/session/selected_camp_provider.dart';
